Final changes for Level-6
